pyproject.toml
[tool.poetry]
name = "alastria-identity-py"
version = "0.1.0"
description = ""
authors = ["Javier Aguirre <javi@wealize.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
web3 = "^5.12.2"
requests = "^2.24.0"
pytest = "^6.1.1"
pyjwt = "^1.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
